cleaned up copy static moving out of main() and using relative path for src and dst

+++ src/main.py
@@ -4,29 +4,36 @@ import os
 import shutil
 
 
-def main():
 
-    source_directory = "/home/madam/workspace/github.com/bootdev/Static-Site-Generator/static"
-    destination_directory = "/home/madam/workspace/github.com/bootdev/Static-Site-Generator/public"
-    #clear destination
-    shutil.rmtree(destination_directory)
-    #remake destination
-    os.mkdir(destination_directory)
-    def copy_static(source_directory, destination_directory):
-        source_list = os.listdir(source_directory)
-        for item in source_list:
-            source_item_path = os.path.join(source_directory, item)
-            destination_item_path = os.path.join(destination_directory, item)
-            #if item is a file that does not exist in destination, copy it
-            if os.path.isfile(source_item_path):
-                shutil.copy(source_item_path, destination_directory)
-            #if item is a directory that does not exist in destination, copy it
-            elif os.path.isdir(source_item_path):
+
+
+def copy_static(source_directory, destination_directory):
+    source_list = os.listdir(source_directory)
+    for item in source_list:
+        source_item_path = os.path.join(source_directory, item)
+        destination_item_path = os.path.join(destination_directory, item)
+        #if item is a file that does not exist in destination, copy it
+        if os.path.isfile(source_item_path):
+            shutil.copy(source_item_path, destination_item_path)
+        #if item is a directory that does not exist in destination, copy it
+        elif os.path.isdir(source_item_path):
+            if not os.path.exists(destination_item_path):
                 os.mkdir(destination_item_path)
-                copy_static(source_item_path, destination_item_path)
+            copy_static(source_item_path, destination_item_path)
 
-    copy_static(source_directory,destination_directory)
-main()
+
+def main():
+    source_directory = "./static"
+    destination_directory = "./public"
+    
+    if os.path.isdir(destination_directory):
+        shutil.rmtree(destination_directory)
+    os.mkdir(destination_directory)
+
+    copy_static(source_directory, destination_directory)
+
+if __name__ == "__main__":
+    main()
